Added a few more functions to mathFunctions.cpp, changed int getFactorial() to unsigned long getFactorial(), commented out FLTK stuff in main.cpp while I figure out how to work with it more.

+++ Calculator/mathFunctions.hpp
@@ -57,7 +57,7 @@ double getTan(double val, bool isRad);
 double getPow(double base, double exp);
 
 //
-int getFactorial(int num);
+unsigned long getFactorial(unsigned long num);
 
 //
 double getFactorial(double num);
